move manifest stuff to manifest.py
